requirements.txt
Django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
psycopg2-binary==2.9.9
mysqlclient==2.2.0
python-decouple==3.8
